backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
python-multipart
PyPDF2
python-docx
openai
pydantic
